Moved old login page to /legacy
